Moved class to util package.
