+++ jbake-core/src/test/java/org/jbake/app/DebugUtilTest.java
@@ -5,6 +5,8 @@ import java.io.PrintStream;
 import java.io.UnsupportedEncodingException;
 import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
+
+import org.jbake.util.DebugUtil;
 import org.junit.Assert;
 import org.junit.Test;
